training_pipeline/requirements.txt
hopsworks[python]==4.1.5
confluent-kafka>=2.3.0
pyarrow>=14.0.1
scikit-learn==1.6.1
xgboost==1.7.3
numpy==1.26.4
pandas==2.1.4
python-dotenv==1.0.0
